requirements.txt
asgiref==3.2.10
autopep8==1.5.3
bootstrap4==0.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.8
django-crispy-forms==1.9.1
django-recaptcha==2.0.6
djangorestframework==3.11.0
idna==2.10
jsonconfig-tool==1.5
keyring==21.2.1
messages==0.5.0
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
python-box==4.2.3
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.1
urllib3==1.25.9
validus==0.3.0
